Update dev requirements and move custom types to the root dir.
