+++ Pipfile
@@ -3,7 +3,7 @@ url = "https://pypi.python.org/simple"
 verify_ssl = true
 
 [dev-packages]
-wemake-python-styleguide = "==0.6.3"
+wemake-python-styleguide = "==0.7.1"
 # see https://github.com/wemake-services/wemake-python-styleguide/pull/472#issuecomment-460057878
 flake8 = '==3.6.0'
 mypy = "==0.670"
@@ -21,7 +21,6 @@ djangorestframework-camel-case = "*"
 django-oauth-toolkit = "*"
 django-cors-middleware = "*"
 pyuploadcare = "*"
-wemake-python-styleguide = "==0.7.1"
 
 [requires]
 python_version = "3.6"
